+++ SkillsTest.Lib/ICourseApi.cs
@@ -6,14 +6,4 @@ namespace SkillsTest.Lib
     {
         Course GetById(int id);
     }
-
-    public class DbCourseApi : ICourseApi
-    {
-        public DataContext Db { get; set; }
-
-        public Course GetById(int id)
-        {
-            return Db.Courses.Where(course => course.Id == id).SingleOrDefault();
-        }
-    }
 }
