+++ SkillsTest.Lib/IStudentApi.cs
@@ -9,14 +9,4 @@ namespace SkillsTest.Lib
     {
         Student GetById(int id);
     }
-
-    public class DbStudentApi : IStudentApi
-    {
-        public DataContext Db { get; set; }
-
-        public Student GetById(int id)
-        {
-            return Db.Students.Where(student => student.Id == id).SingleOrDefault();
-        }
-    }
 }
